Add interface for defining columns of nested collections
