+++ src/realm/spec.hpp
@@ -64,6 +64,9 @@ public:
     ColumnAttrMask get_column_attr(size_t column_ndx) const noexcept;
     void set_dictionary_key_type(size_t column_ndx, DataType key_type);
     DataType get_dictionary_key_type(size_t column_ndx) const;
+    void set_nested_column_types(size_t column_ndx, const std::vector<CollectionType>& types);
+    CollectionType get_nested_column_type(size_t column_ndx, size_t level) const;
+    size_t get_nesting_levels(size_t column_ndx) const;
 
     // Auto Enumerated string columns
     void upgrade_string_to_enum(size_t column_ndx, ref_type keys_ref);
@@ -88,14 +91,22 @@ public:
 private:
     // Underlying array structure.
     //
+    static constexpr size_t s_types_ndx = 0;
+    static constexpr size_t s_names_ndx = 1;
+    static constexpr size_t s_attributes_ndx = 2;
+    static constexpr size_t s_nested_types_ndx = 3;
+    static constexpr size_t s_enum_keys_ndx = 4;
+    static constexpr size_t s_col_keys_ndx = 5;
+    static constexpr size_t s_spec_max_size = 6;
+
     Array m_top;
     Array m_types;            // 1st slot in m_top
     ArrayStringShort m_names; // 2nd slot in m_top
     Array m_attr;             // 3rd slot in m_top
-    // 4th slot in m_top is vacant
+                              // 4th slot in m_top not cached
     Array m_enumkeys; // 5th slot in m_top
     Array m_keys;     // 6th slot in m_top
-    size_t m_num_public_columns;
+    size_t m_num_public_columns = 0;
 
     Spec(Allocator&) noexcept; // Unattached
 
@@ -123,18 +134,6 @@ private:
     bool convert_column_attributes();
     bool convert_column_keys(TableKey table_key);
     void fix_column_keys(TableKey table_key);
-    bool has_subspec()
-    {
-        return m_top.get(3) != 0;
-    }
-    void destroy_subspec()
-    {
-        Node::destroy(m_top.get_as_ref(3), m_top.get_alloc());
-        m_top.set(3, 0);
-    }
-    TableKey get_opposite_link_table_key(size_t column_ndx) const noexcept;
-    size_t get_origin_column_ndx(size_t backlink_col_ndx) const noexcept;
-    ColKey find_backlink_column(TableKey origin_table_key, size_t spec_ndx) const noexcept;
 
 
     // Generate a column key only from state in the spec.
@@ -165,11 +164,11 @@ inline Spec::Spec(Allocator& alloc) noexcept
     , m_enumkeys(alloc)
     , m_keys(alloc)
 {
-    m_types.set_parent(&m_top, 0);
-    m_names.set_parent(&m_top, 1);
-    m_attr.set_parent(&m_top, 2);
-    m_enumkeys.set_parent(&m_top, 4);
-    m_keys.set_parent(&m_top, 5);
+    m_types.set_parent(&m_top, s_types_ndx);
+    m_names.set_parent(&m_top, s_names_ndx);
+    m_attr.set_parent(&m_top, s_attributes_ndx);
+    m_enumkeys.set_parent(&m_top, s_enum_keys_ndx);
+    m_keys.set_parent(&m_top, s_col_keys_ndx);
 }
 
 inline bool Spec::init_from_parent() noexcept
